Add bake support (#11)
resolves #6

+++ internal/cacher/cacher.go
@@ -3,93 +3,109 @@ package cacher
 import (
 	"encoding/json"
 	"errors"
+	"log/slog"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/elliotchance/orderedmap/v3"
-	log "github.com/sirupsen/logrus"
+	"github.com/samber/lo"
 
 	"github.com/apparentlymart/go-userdirs/userdirs"
-	"github.com/hytromo/mimosa/internal/docker"
 	"github.com/hytromo/mimosa/internal/hasher"
+	"github.com/hytromo/mimosa/internal/utils/envutil"
 	"github.com/hytromo/mimosa/internal/utils/fileutil"
 )
 
-var CacheDir = userdirs.ForApp("mimosa", "hytromo", "mimosa.hytromo.com").CacheDir
+var CacheDir = envutil.GetEnv("MIMOSA_CACHE_DIR", userdirs.ForApp("mimosa", "hytromo", "mimosa.hytromo.com").CacheDir)
 
 type CacheFile struct {
-	Tags          []string  `json:"tags"`
-	LastUpdatedAt time.Time `json:"lastUpdatedAt"`
+	TagsByTarget  map[string][]string `json:"tagsByTarget"`
+	LastUpdatedAt time.Time           `json:"lastUpdatedAt"`
 }
 
+// Cache represents the final hash of the currently running command and files
+// and the current available in-memory cache entries
 type Cache struct {
-	InMemoryEntries *orderedmap.OrderedMap[string, string] // populated by the MIMOSA_CACHE environment variable and taking precedence over the cache directory
-	FinalHash       string                                 // the final hash of the current command and files
+	InMemoryEntries *InMemoryCache // populated by the "InjectCacheEnvVarName" environment variable and taking precedence over the cache directory
+	CacheDir        string         // the directory where the cache files are stored - defaults to CacheDir
+	Hash            string         // the final hash of the current command and files
 }
 
 func (cache *Cache) DataPath() string {
-	return filepath.Join(CacheDir, cache.FinalHash+".json")
+	return filepath.Join(cache.CacheDir, cache.Hash+".json")
+}
+
+func (cache *Cache) GetLatestTagByTarget() (map[string]string, error) {
+	var tagsByTarget map[string][]string
+
+	inMemoryEntry, exists := cache.GetInMemoryEntry()
+
+	if exists {
+		tagsByTarget = inMemoryEntry.TagsByTarget
+	} else {
+		data, err := os.ReadFile(cache.DataPath())
+		if err != nil {
+			return nil, err
+		}
+
+		var cacheFile CacheFile
+		err = json.Unmarshal(data, &cacheFile)
+		if err != nil {
+			return nil, err
+		}
+
+		tagsByTarget = cacheFile.TagsByTarget
+	}
+
+	latestTagByTarget := make(map[string]string)
+
+	for target, tags := range tagsByTarget {
+		if len(tags) > 0 {
+			// for each of the targets get the most recent cached tag
+			latestTagByTarget[target] = tags[len(tags)-1]
+		}
+	}
+
+	return latestTagByTarget, nil
+}
+
+func (cache *Cache) ExistsInFilesystem() bool {
+	if _, err := os.Stat(cache.DataPath()); errors.Is(err, os.ErrNotExist) {
+		return false
+	}
+	return true
 }
 
 func (cache *Cache) Remove(dryRun bool) error {
 	if dryRun {
-		log.Infoln("> DRY RUN: cache entry would be removed from", cache.DataPath())
+		slog.Info("> DRY RUN: cache entry would be removed from", "path", cache.DataPath())
 		return nil
 	}
 
 	return os.Remove(cache.DataPath())
 }
 
-func (cache *Cache) LatestTag() (string, error) {
-	inMemoryEntry, ok := cache.GetInMemoryEntry()
-
-	if ok {
-		log.Debugf("Returning in-memory cache entry for hash %s: %s", cache.FinalHash, inMemoryEntry)
-		return inMemoryEntry, nil
+func (cache *Cache) GetInMemoryEntry() (CacheFile, bool) {
+	if cache.InMemoryEntries == nil || cache.InMemoryEntries.Len() == 0 {
+		return CacheFile{}, false
 	}
 
-	cachedFilePath := cache.DataPath()
-
-	data, err := os.ReadFile(cachedFilePath)
+	z85Hash, err := hasher.HexToZ85(cache.Hash)
 	if err != nil {
-		return "", err
-	}
-
-	var cacheFile CacheFile
-	err = json.Unmarshal(data, &cacheFile)
-	if err != nil {
-		return "", err
-	}
-	if len(cacheFile.Tags) == 0 {
-		return "", nil
-	}
-
-	return cacheFile.Tags[len(cacheFile.Tags)-1], nil
-}
-
-func (cache *Cache) GetInMemoryEntry() (string, bool) {
-	if cache.InMemoryEntries.Len() == 0 {
-		return "", false
-	}
-
-	// first checking the in-memory cache
-	z85Hash, err := hasher.HexToZ85(cache.FinalHash)
-	if err != nil {
-		log.Warnf("Failed to convert final hash to Z85: %v", err)
-		return "", false
+		slog.Warn("Failed to convert final hash to Z85", "error", err)
+		return CacheFile{}, false
 	}
 	if entry, ok := cache.InMemoryEntries.Get(z85Hash); ok {
 		return entry, true
 	}
 
-	return "", false
+	return CacheFile{}, false
 }
 
 func (cache *Cache) Exists() bool {
 	if _, ok := cache.GetInMemoryEntry(); ok {
-		log.Debugf("Cache hit in memory for hash %s", cache.FinalHash)
+		slog.Debug("Cache hit in memory for hash", "hash", cache.Hash)
 		return true
 	}
 
@@ -97,116 +113,64 @@ func (cache *Cache) Exists() bool {
 		return false
 	}
 
-	log.Debugf("Cache hit on disk for hash %s", cache.FinalHash)
+	slog.Debug("Cache hit on disk for hash", "hash", cache.Hash)
 
 	return true
 }
 
-func GetCache(parsedBuildCommand docker.ParsedBuildCommand) (cache Cache, err error) {
-	commandInfo := make([]string, len(parsedBuildCommand.CmdWithTagPlaceholder))
-	copy(commandInfo, parsedBuildCommand.CmdWithTagPlaceholder)
-
-	commandInfo = append(commandInfo, parsedBuildCommand.RegistryDomain)
-
-	commandHash := hasher.HashStrings(commandInfo)
-
-	log.Debugf("Command hash: %v - deducted from %v", commandHash, commandInfo)
-
-	if _, err := os.Stat(parsedBuildCommand.ContextPath); errors.Is(err, os.ErrNotExist) {
-		return cache, err
-	}
-
-	if _, err := os.Stat(parsedBuildCommand.DockerfilePath); errors.Is(err, os.ErrNotExist) {
-		return cache, err
-	}
-
-	if parsedBuildCommand.DockerignorePath != "" {
-		if _, err := os.Stat(parsedBuildCommand.DockerignorePath); errors.Is(err, os.ErrNotExist) {
-			return cache, err
-		}
-	}
-
-	files, err := docker.IncludedFiles(parsedBuildCommand.ContextPath, parsedBuildCommand.DockerignorePath)
-
-	if err != nil {
-		return cache, err
-	}
-
-	if parsedBuildCommand.DockerignorePath != "" {
-		files = append(files, parsedBuildCommand.DockerignorePath)
-	}
-
-	if parsedBuildCommand.DockerfilePath != "" {
-		files = append(files, parsedBuildCommand.DockerfilePath)
-	}
-
-	filesHash, err := hasher.HashFiles(files)
-
-	log.Debugf("Files hash: %v - deducted from %v files", filesHash, len(files))
-
-	if err != nil {
-		return cache, err
-	}
-
-	cache.FinalHash = hasher.HashStrings([]string{commandHash, filesHash})
-
-	log.Debugf("Final hash of command and files: %v", cache.FinalHash)
-
-	cache.InMemoryEntries = GetAllInMemoryEntries()
-
-	return cache, nil
-}
-
-func (cache *Cache) Save(finalTag string, dryRun bool) (dataFile string, err error) {
-	dataFile = cache.DataPath()
+func (cache *Cache) Save(tagsByTarget map[string][]string, dryRun bool) error {
+	dataFile := cache.DataPath()
 
 	if dryRun {
-		log.Infoln("> DRY RUN: cache entry would be saved to", dataFile)
-		return dataFile, nil
+		slog.Info("> DRY RUN: cache entry would be saved to", "path", dataFile)
+		return nil
 	}
 
 	if err := os.MkdirAll(filepath.Dir(dataFile), 0755); err != nil {
-		return dataFile, err
+		return err
 	}
 
-	tags := make([]string, 0, 10)
-	tagSet := make(map[string]struct{})
+	var td CacheFile
 
-	// Read existing tags from mimosa-cache.json if it exists
+	// Read existing tags from the cache file if it exists
 	if data, err := os.ReadFile(dataFile); err == nil {
-		var td CacheFile
-		if err := json.Unmarshal(data, &td); err == nil {
-			for _, tag := range td.Tags {
-				if tag == "" {
-					continue
-				}
-				if _, exists := tagSet[tag]; !exists {
-					tags = append(tags, tag)
-					tagSet[tag] = struct{}{}
-				}
-			}
+		if err := json.Unmarshal(data, &td); err != nil {
+			slog.Debug("Failed to unmarshal cache file", "path", dataFile, "error", err)
 		}
 	}
 
-	// Add new tag if unique at the end
-	if _, exists := tagSet[finalTag]; !exists {
-		tags = append(tags, finalTag)
-	}
-	// Keep only last 10 unique tags (= the last 10 tags of the list)
-	if len(tags) > 10 {
-		tags = tags[len(tags)-10:]
+	if td.TagsByTarget == nil {
+		td.TagsByTarget = make(map[string][]string)
 	}
 
-	return dataFile, fileutil.SaveJSON(dataFile, CacheFile{
-		Tags:          tags,
-		LastUpdatedAt: time.Now().UTC(),
-	})
+	// add the new tags to the existing tags
+	for target, tags := range tagsByTarget {
+		for _, tag := range tags {
+			if _, exists := td.TagsByTarget[target]; !exists {
+				td.TagsByTarget[target] = []string{tag}
+			} else {
+				td.TagsByTarget[target] = append(td.TagsByTarget[target], tag)
+			}
+
+			// keep at most 10 tags per target
+			if len(td.TagsByTarget[target]) > 10 {
+				td.TagsByTarget[target] = td.TagsByTarget[target][len(td.TagsByTarget[target])-10:]
+			}
+		}
+		td.TagsByTarget[target] = lo.Uniq(td.TagsByTarget[target])
+	}
+
+	td.LastUpdatedAt = time.Now().UTC()
+
+	return fileutil.SaveJSON(dataFile, td)
 }
 
-func ForgetCacheEntriesOlderThan(forgetTime time.Time) {
-	cacheDir := CacheDir
+func ForgetCacheEntriesOlderThan(forgetTime time.Time, cacheDir string, dryRun bool) error {
+	slog.Debug("Forgetting cache entries older than", "forgetTime", forgetTime, "cacheDir", cacheDir)
 
-	log.Debugf("Forgetting cache entries older than %s in %s", forgetTime, cacheDir)
+	if _, err := os.Stat(cacheDir); errors.Is(err, os.ErrNotExist) {
+		return nil
+	}
 
 	deletedCount := 0
 	err := filepath.Walk(cacheDir, func(path string, info os.FileInfo, err error) error {
@@ -223,32 +187,34 @@ func ForgetCacheEntriesOlderThan(forgetTime time.Time) {
 			return err
 		}
 
-		log.Debugln("Checking cache file:", path)
+		slog.Debug("Checking cache file", "path", path)
 
 		var cacheFile CacheFile
 		if err := json.Unmarshal(data, &cacheFile); err != nil {
-			log.Errorf("Failed to unmarshal cache file %s: %v", path, err)
+			slog.Error("Failed to unmarshal cache file", "path", path, "error", err)
 			return nil
 		}
 
 		if cacheFile.LastUpdatedAt.After(forgetTime) {
-			log.Debugf("Cache file %s is newer than forget time, skipping deletion", path)
+			slog.Debug("Cache file is newer than forget time, skipping deletion", "path", path)
 			return nil
 		}
 
-		log.Debugf("Cache file %s is older than forget time, deleting", path)
-		if err := os.Remove(path); err != nil {
-			log.Errorf("Failed to delete cache file %s: %v", path, err)
-			return nil
+		slog.Debug("Cache file is older than forget time, deleting", "path", path)
+		if dryRun {
+			slog.Info("File not deleted due to dry run", "path", path)
+		} else {
+			if err := os.Remove(path); err != nil {
+				slog.Error("Failed to delete cache file", "path", path, "error", err)
+				return nil
+			}
 		}
 
 		deletedCount++
 		return nil
 	})
 
-	if err != nil {
-		log.Errorf("Failed to forget cache entries: %v", err)
-	}
+	slog.Info("Deleted cache entries older than", "count", deletedCount, "forgetTime", forgetTime)
 
-	log.Infoln("Deleted", deletedCount, "cache entries older than", forgetTime)
+	return err
 }
